requirements.txt
absl-py==2.1.0
alfred==0.3
alfred-py==2.10.1
asttokens==2.4.1
astunparse==1.6.3
babel==2.16.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
Deprecated==1.2.14
executing==2.1.0
filelock==3.16.1
Flask==2.2.5
Flask-Babel==2.0.0
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
funcy==2.0
future==1.0.0
gast==0.4.0
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.66.1
gtrending==0.5.1
h5py==3.11.0
icecream==2.1.3
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsons==1.6.3
keras==2.13.1
kiwisolver==1.4.7
libclang==18.1.1
loguru==0.7.2
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
natsort==8.4.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.1
pandas==2.0.3
pascal-voc-writer==0.1.4
pillow==10.4.0
pip==24.2
portalocker==2.10.1
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.7
Pygments==2.18.0
pyparsing==3.1.4
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==306
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.1
rsa==4.9
setuptools==75.1.0
six==1.16.0
soupsieve==2.6
sympy==1.13.3
tbb==2021.13.1
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typing_extensions==4.12.2
typish==1.9.3
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
wheel==0.44.0
win32-setctime==1.1.0
wrapt==1.16.0
xdg==6.0.0
zipp==3.20.2
